tools/requirements.txt
openapi-python-client
pyyaml
